requirements.txt
asgiref==3.8.1
astor==0.8.1
attrs==23.2.0
bandit==1.7.9
darglint==1.8.1
Django==5.0.7
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
docutils==0.21.2
eradicate==2.3.0
flake8==7.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==24.4.26
flake8-commas==2.1.0
flake8-comprehensions==3.15.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-polyfill==1.0.2
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gitdb==4.0.11
GitPython==3.1.43
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
pbr==6.0.0
pep8-naming==0.13.3
phonenumbers==8.13.42
psycopg==3.1.18
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-usernames==1.0.0
PyYAML==6.0.1
redgreenunittest==0.1.1
restructuredtext_lint==1.4.0
rich==13.7.1
setuptools==71.1.0
smmap==5.0.1
snowballstemmer==2.2.0
sqlparse==0.5.0
stevedore==5.2.0
typing_extensions==4.12.2
wemake-python-styleguide==0.19.2
